pyproject.toml
[project]
name = "writher-linux"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "customtkinter>=5.2.0",
    "pynput>=1.7.6",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "httpx>=0.27.0",
    "pyperclip>=1.8.2",
    "dbus-python>=1.3.2",
    "PyGObject>=3.42.0",
]

[project.scripts]
writher = "main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
